+++ KachaowAuto/KachaowAuto/Controllers/PartCategoryController.cs
@@ -1,5 +1,6 @@
 ﻿using KachaowAuto.Data;
 using KachaowAuto.Data.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -7,6 +8,7 @@ using System.IO;
 
 namespace KachaowAuto.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class PartCategoryController : Controller
     {
         private readonly KachaowAutoDbContext context;
@@ -14,7 +16,8 @@ namespace KachaowAuto.Controllers
         {
             context = _context;
         }
-        // GET: PartCategoryController
+
+        [Authorize(Roles = "Admin,Mechanic")]
         public async Task<IActionResult> Index()
         {
             var partCategories = await context.PartCategories
@@ -23,14 +26,12 @@ namespace KachaowAuto.Controllers
             return View(partCategories);
         }
 
-        // GET: PartCategoryController/Create
         public async Task<IActionResult> Create()
         {
             ViewBag.Parts = await context.Parts.ToListAsync();
             return View();
         }
 
-        // POST: PartCategoryController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(PartCategory partCategory)
@@ -45,7 +46,6 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: PartCategoryController/Edit/5
         public async Task<IActionResult> Edit(int id)
         {
             ViewBag.Parts = await context.Parts.ToListAsync();
@@ -57,7 +57,6 @@ namespace KachaowAuto.Controllers
             return View(partCategory);
         }
 
-        // POST: PartCategoryController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(PartCategory partCategory)
@@ -74,7 +73,6 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: PartCategoryController/Delete/5
         public async Task<IActionResult> Delete(int id)
         {
             var partCategory = await context.PartCategories.FirstOrDefaultAsync(a => a.PartCategoryId == id);
@@ -85,12 +83,16 @@ namespace KachaowAuto.Controllers
             return View(partCategory);
         }
 
-        // POST: PartCategoryController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var partCategory = await context.PartCategories.FirstOrDefaultAsync(a => a.PartCategoryId == id);
+
+            if (partCategory == null)
+            {
+                return NotFound();
+            }
 
             context.PartCategories.Remove(partCategory);
             await context.SaveChangesAsync();
