+++ KachaowAuto/KachaowAuto/Controllers/PartController.cs
@@ -1,11 +1,13 @@
 ﻿using KachaowAuto.Data;
 using KachaowAuto.Data.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace KachaowAuto.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class PartController : Controller
     {
         private readonly KachaowAutoDbContext context;
@@ -13,7 +15,8 @@ namespace KachaowAuto.Controllers
         {
             context = _context;
         }
-        // GET: PartController
+
+        [Authorize(Roles = "Admin,Mechanic")]
         public async Task<IActionResult> Index()
         {
             var parts = await context.Parts
@@ -24,24 +27,21 @@ namespace KachaowAuto.Controllers
             return View(parts);
         }
 
-        // GET: PartController/Create
         public async Task<IActionResult> Create()
         {
-            ViewBag.Brands = await context.Brands.ToListAsync();
+            ViewBag.Categories = await context.PartCategories.ToListAsync();
             ViewBag.AppointmentParts = await context.AppointmentParts.ToListAsync();
             ViewBag.Images = await context.PartImages.ToListAsync();
             return View();
         }
 
-
-        // POST: PartController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Part part)
         {
             if (!ModelState.IsValid)
             {
-                ViewBag.Brands = await context.Brands.ToListAsync();
+                ViewBag.Categories = await context.PartCategories.ToListAsync();
                 ViewBag.AppointmentParts = await context.AppointmentParts.ToListAsync();
                 ViewBag.Images = await context.PartImages.ToListAsync();
                 return View(part);
@@ -51,10 +51,9 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: PartController/Edit/5
         public async Task<IActionResult> Edit(int id)
         {
-            ViewBag.Brands = await context.Brands.ToListAsync();
+            ViewBag.Categories = await context.PartCategories.ToListAsync();
             ViewBag.AppointmentParts = await context.AppointmentParts.ToListAsync();
             ViewBag.Images = await context.PartImages.ToListAsync();
             var part = await context.Parts.FirstOrDefaultAsync(a => a.PartId == id);
@@ -65,14 +64,13 @@ namespace KachaowAuto.Controllers
             return View(part);
         }
 
-        // POST: PartController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Part part)
         {
             if (!ModelState.IsValid)
             {
-                ViewBag.Brands = await context.Brands.ToListAsync();
+                ViewBag.Categories = await context.PartCategories.ToListAsync();
                 ViewBag.AppointmentParts = await context.AppointmentParts.ToListAsync();
                 ViewBag.Images = await context.PartImages.ToListAsync();
                 return View(part);
@@ -84,7 +82,6 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: PartController/Delete/5
         public async Task<IActionResult> Delete(int id)
         {
             var part = await context.Parts.FirstOrDefaultAsync(a => a.PartId == id);
@@ -95,12 +92,16 @@ namespace KachaowAuto.Controllers
             return View(part);
         }
 
-        // POST: PartController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var part = await context.Parts.FirstOrDefaultAsync(a => a.PartId == id);
+
+            if (part == null)
+            {
+                return NotFound();
+            }
 
             context.Parts.Remove(part);
             await context.SaveChangesAsync();
