+++ KachaowAuto/KachaowAuto/Controllers/RegionController.cs
@@ -1,11 +1,13 @@
 ﻿using KachaowAuto.Data;
 using KachaowAuto.Data.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace KachaowAuto.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class RegionController : Controller
     {
         private readonly KachaowAutoDbContext context;
@@ -13,7 +15,8 @@ namespace KachaowAuto.Controllers
         {
             context = _context;
         }
-        // GET: RegionController
+
+        [Authorize(Roles = "Admin,Mechanic,Client")]
         public async Task<IActionResult> Index()
         {
             var regions = await context.Regions
@@ -22,14 +25,12 @@ namespace KachaowAuto.Controllers
             return View(regions);
         }
 
-        // GET: RegionController/Create
         public async Task<IActionResult> Create()
         {
             ViewBag.Workshops = await context.Workshops.ToListAsync();
             return View();
         }
 
-        // POST: RegionController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Region region)
@@ -44,7 +45,6 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: RegionController/Edit/5
         public async Task<IActionResult> Edit(int id)
         {
             ViewBag.Workshops = await context.Workshops.ToListAsync();
@@ -56,7 +56,6 @@ namespace KachaowAuto.Controllers
             return View(region);
         }
 
-        // POST: RegionController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Region region)
@@ -73,7 +72,6 @@ namespace KachaowAuto.Controllers
             return RedirectToAction("Index");
         }
 
-        // GET: RegionController/Delete/5
         public async Task<IActionResult> Delete(int id)
         {
             var region = await context.Regions.FirstOrDefaultAsync(a => a.RegionId == id);
@@ -84,12 +82,16 @@ namespace KachaowAuto.Controllers
             return View(region);
         }
 
-        // POST: RegionController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var region = await context.Regions.FirstOrDefaultAsync(a => a.RegionId == id);
+
+            if (region == null)
+            {
+                return NotFound();
+            }
 
             context.Regions.Remove(region);
             await context.SaveChangesAsync();
